moved function loader to module
